Move `logger` module into its own file
